requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pefile==2023.2.7
r2pipe==1.9.0
google-genai==0.3.0
python-multipart==0.0.9
